+++ lite/appspawn_service.c
@@ -111,7 +111,7 @@ static int GetMessageSt(MessageSt *msgSt, IpcIo *req)
 }
 
 static AppSpawnContentLite *g_appSpawnContentLite = NULL;
-AppSpawnContent *AppSpawnCreateContent(const char *socketName, char *longProcName, int64_t longProcNameLen, int cold)
+AppSpawnContent *AppSpawnCreateContent(const char *socketName, char *longProcName, uint32_t longProcNameLen, int cold)
 {
     UNUSED(longProcName);
     UNUSED(longProcNameLen);
